requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==5.0.4
django-admin-sortable==2.3
django-admin-sortable2==2.2.3
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-multiupload==0.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.8
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.32.3
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
